requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
PyJWT==2.8.1
passlib==1.7.4
bcrypt==4.1.2
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==25.2.0
eventlet==0.36.3
